# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Temporary Group Maker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (for member search)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /groups/{groupId}/summary:
    get:
      operationId: getGroupSummary
      tags: [summaries]
      summary: Get the AI-generated summary for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupSummary"
    post:
      operationId: generateGroupSummary
      tags: [summaries]
      summary: Trigger AI summary generation for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupSummary"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          nullable: true
      required:
        - name
        - email
...
    GroupSummary:
      type: object
      properties:
        id:
          type: string
        groupId:
          type: string
        generatedAt:
          type: string
          format: date-time
        keyDecisions:
          type: array
          items:
            type: string
        actionItems:
          type: array
          items:
            type: string
        participantCount:
          type: integer
        suggestedNextAction:
          type: string
          enum: [archive, extend, branch, task_board, decision_board]
          nullable: true
      required:
        - id
        - groupId
        - generatedAt
        - keyDecisions
        - actionItems
        - participantCount
...