# lib/api-zod/src/generated/types/groupSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Temporary Group Maker API
 * OpenAPI spec version: 0.1.0
 */
import type { GroupSummarySuggestedNextAction } from "./groupSummarySuggestedNextAction";

export interface GroupSummary {
  id: string;
  groupId: string;
  generatedAt: Date;
  keyDecisions: string[];
  actionItems: string[];
  participantCount: number;
  suggestedNextAction?: GroupSummarySuggestedNextAction;
}